backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0  # For future PostgreSQL migration
greenlet>=3.0.0  # Required for async SQLAlchemy

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for passlib compatibility
cryptography==41.0.7

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/LLM Integration
google-cloud-aiplatform>=1.38.1
httpx==0.26.0  # For local LLM API calls

# Health Data Integration
requests==2.31.0

# Date/Time handling
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
